requirements.txt
numpy>=1.20.0
matplotlib>=3.5.0
pyserial>=3.5
ikpy>=3.2.0 